cybellum-task/backend/blog/requirements.txt
Flask
SQLAlchemy
fastapi
psycopg2-binary==2.9.1
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
Flask-WTF
Werkzeug==2.0.1
passlib==1.7.4
python-jose
uvicorn